requirements.txt
numpy==1.18.5
opencv==4.4.0.44
streamlit==0.67.1
tensorflow==2.3.1
matplotlib==3.3.2
pillow==7.2.0
nbimporter==0.3.3